Fix to errant selectBoxList closing
For taller SelectBoxes than ones that are created with the default skin (tested on 43 pixels tall), on its update the condition governing whether the list should hide again was fulfilled when it shouldn't have been, resulting in the SelectBox immediately hiding its list after opening if you clicked more than halfway up. On a 43-px tall one, that's pretty much the magic 20 that was here.