Group x/y fix
Groups could show unexpected behaviour if the x/y changed during the draw calls. This could be seen with two dropdowns side by side - click one dropdown to open it, and then next click the other dropdown, and the second list would be shown in the same place as the first was.
This change keeps any changes from draw calls, but still removes the offset. This seems to have resolved the issue.