Load particle texture from directory
Allows loading of particle texture from the directory the .pex file is in.
In the .pex file, the data attribute doesn't need to be used - as long as you specify the name of a texture relative to the .pex file, though it definitely doesn't support going up levels. 
For example - <texture name="particle.png" /> is perfectly valid, as long as there's a texture.png file in the same directory as the .pex file.
The method of using <texture name="name" data="TIFF->base64->gzip" /> is still usable