In ProgressBar, changed .setValue() by adding option to ignore snap and reversing condition.
Personally, I believe that it's easier to understand the meaning behind this condition when conditions are not negated.

Additionally I thought that it's useful to manually set the value to anything we want (within min/max range) while ignoring snap - I added another parameter to .setValue() (defaults to false) that, when true, will not snap the value.