Do not round value when value is lower/higher than min/max.
Previously, when using stepValue that's really close to max, but not equal to it - we would get a hard time actually getting that value because of rounding.

With this change we no longer call rounding code when we're lower or higher than min or max and instead, we set value to either min or max.