Default ParticleEmitterConfig Colors 1:1 Sprite
Changed the defaults for the color fields inside the config to have no variance at all and to just output the particle with no fading or random colors.

StartColorVariance was (255,255,255,255) so a random color was picked by default, now its (0,0,0,0)
FinishColor was (0,0,0,255) so it would turn slowly turn black.
FinishColorVariance was (0,0,0,255) so a random alpha was picked by default, now its (0,0,0,0)