Add JustPushed and repeating support
VirtualIntegerAxis now has the JustPushed method and DirectionJustPushed property, which can be used to check if any of its nodes were just pushed in the current frame (similar to VirtualButton.IsPressed). It also adds repeating input support, which allow JustPushed to repeatedly fire when the parameters are set with the new SetRepeat() methods. This functions the same as VirtualButton's SetRepeat feature.