Add JustPushed and SetRepeat methods to VirtualIntegerAxis (#756)
* Add axis JustPushed support

VirtualAxis nodes now have the `bool JustPushed(int direction)` method and `int DirectionJustPushed` property. These can be used to check if the axis was just pushed in a certain direction in the current frame, similar to `VirtualButton.IsPressed`.

* Add JustPushed and repeating support

VirtualIntegerAxis now has the JustPushed method and DirectionJustPushed property, which can be used to check if any of its nodes were just pushed in the current frame (similar to VirtualButton.IsPressed). It also adds repeating input support, which allow JustPushed to repeatedly fire when the parameters are set with the new SetRepeat() methods. This functions the same as VirtualButton's SetRepeat feature.